requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1.2
djangorestframework==3.9.0
httpie==0.9.9
idna==2.7
pkg-resources==0.0.0
Pygments==2.2.0
pytz==2018.7
requests==2.20.0
urllib3==1.24
